Add color deficiency types and random quote generator; implement filters for contrast simulation
