+++ src/components/apps/contrast-checker/controls/Controls.scss
@@ -1,0 +1,174 @@
+.cxc-main {
+    &__table {
+        tr {
+            &:not(:last-child) {
+                border: var(--_border);
+            }
+        }
+        th, 
+        td {
+            padding: 0.5rem;
+            text-align: center;
+            line-height: 1;
+            &:not(:last-child) {
+                border-right: var(--_border);
+            }
+            &:is(:first-child) {
+                text-align: left;
+            }
+            &:is(td) {
+                &:is(:nth-last-child(1), :nth-last-child(2)) {
+                    &:before {
+                        --_color: #ff0000;
+                        content: '';
+                        background-color: var(--_color);
+                        mask: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='24' height='24' viewBox='0 0 24 24' fill='none' stroke='currentColor' stroke-width='1.25' stroke-linecap='round' stroke-linejoin='round'%3E%3Cpath d='M18 6 6 18'/%3E%3Cpath d='m6 6 12 12'/%3E%3C/svg%3E") no-repeat center;
+                        height: 1rem;
+                        width: 1rem;
+                        mask-size: contain;
+                        display: inline-block;
+                        margin-right: 0.5rem;
+                        translate: 0 2.5px;
+                    }
+                    &[data-pass='true'] {
+                        &::before {
+                            --_color: light-dark(green, lime);
+                            mask-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='24' height='24' viewBox='0 0 24 24' fill='none' stroke='currentColor' stroke-width='1.25' stroke-linecap='round' stroke-linejoin='round'%3E%3Cpath d='M20 6 9 17l-5-5'/%3E%3C/svg%3E");
+                        }
+                    }
+                }
+            }
+        }
+        & {
+            --_border: 1px solid var(--app-color-border-primary);
+            width: 100%;
+            border-collapse: collapse;
+            border: var(--_border);
+        }
+        &wrap {
+            padding: var(--_padding);
+        } 
+    }
+    &__setup {
+
+        &-field {
+            > input[type='color'] {
+                // appearance: none;
+                height: var(--_size);
+                width: var(--_size);
+                border-radius: var(--_child-radius);
+                min-width: 0;
+            }
+            > p {
+                height: var(--_size);
+                display: flex;
+                flex-direction: column;
+                font-size: 1.125rem;
+                border-radius: var(--_child-radius);
+                background-color: light-dark(#eee, #575757);
+                padding: 0.625rem;
+                line-height: 1;
+                gap: 0.5rem;
+                text-transform: uppercase;
+                &::before {
+                    content: 'HEX';
+                    font-size: 0.75rem;
+                }
+            }
+            > legend {
+                grid-column: 1/-1;
+                float: inline-start;
+            }
+            & {
+                --_child-radius: var(--_radius);
+                --_size: 3.375rem;
+                display: grid;
+                grid-template-columns: 1fr var(--_size);
+                gap: 0.75rem 0.625rem;
+            }
+        }
+        &-body {
+            display: grid;
+            grid-template-columns: repeat(2, 1fr);
+            gap: 1rem 2rem;
+            padding: var(--_padding);
+            @media (min-width: 700px) and (max-width: 1200px) {
+                grid-template-columns: 1fr;
+            }
+        }
+        &-head {
+            > p {
+                span {
+                    font-weight: 500;
+                    padding: 0.375rem;
+                    border-radius: var(--_radius);
+                    background-color: var(--_bg);
+                }
+                --_bg: light-dark(#eee, #575757);
+                font-size: 1.25rem;
+                &[data-clears='aa'] {
+                    --_bg: light-dark(#f2c59c, #ff7b07);
+                }
+                &[data-clears='aaa'] {
+                    --_bg: light-dark(#9bea9b, #189d18);
+                }
+
+            }
+            > button {
+                &::before {
+                    content: '';
+                    height: 1.5rem;
+                    width: 1.5rem;
+                    mask: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='24' height='24' fill='none' viewBox='0 0 24 24' aria-hidden='true'%3E%3Cpath fill='currentColor' fill-opacity='0.9' fill-rule='evenodd' d='m17.292 7.998-.146-.146-1-1a.5.5 0 0 1 .708-.708l1 1 1 1a.5.5 0 0 1 0 .708l-2 2a.5.5 0 0 1-.708-.708L17.292 9h-1.348a3.5 3.5 0 0 0-2.752 1.337l-3.097 3.943a4.5 4.5 0 0 1-3.54 1.72H5.5a.5.5 0 1 1 0-1.001h1.056a3.5 3.5 0 0 0 2.752-1.338l2.308-2.937.134-.172.655-.833a4.5 4.5 0 0 1 3.54-1.72zm-9.438 2.854a.5.5 0 0 0 0-.708L6.707 8.998h.849a3.5 3.5 0 0 1 2.237.809l.617-.79a4.5 4.5 0 0 0-2.854-1.02h-.847l.145-.145 1-1a.5.5 0 0 0-.708-.708l-1 1-1 1a.5.5 0 0 0 0 .708l2 2a.5.5 0 0 0 .708 0m5.551 3.426-.383-.488.634-.812.536.682a3.5 3.5 0 0 0 2.752 1.337H18a.5.5 0 1 1 0 1.001h-1.056a4.5 4.5 0 0 1-3.539-1.72' clip-rule='evenodd'%3E%3C/path%3E%3C/svg%3E") no-repeat center;
+                    mask-size: contain;
+                    background-color: var(--_color);
+                }
+                --_color: var(--color-default);
+                display: flex;
+                background-color: transparent;
+                align-items: center;
+                padding: 0;
+                border: none;
+                gap: 0.5rem;
+            }
+            & {
+                display: flex;
+                justify-content: space-between;
+                align-items: center;
+                padding: var(--_padding);
+                border: 1px solid var(--app-color-border-secondary);
+                border-left: none;
+                border-right: none;
+            }
+        }
+        &-peak {
+            > div {
+                &:first-child {
+                    border-top-left-radius: var(--_radius);
+                    background-color: var(--_background, #fff);
+                }
+                &:last-child {
+                    border-top-right-radius: var(--_radius);
+                    background-color: var(--_foreground, #000);
+                }
+                & {
+                    height: var(--_padding, 1.5rem);
+                }
+            }
+            & {
+                display: grid;
+                grid-template-columns: 1fr 1fr;
+            }
+        }
+        & {
+            --_radius: 0.5rem;
+            border: 1px solid var(--app-color-border-secondary);
+            border-radius: var(--_radius);
+            // box-shadow: rgba(0, 0, 0, 0.1) 0px 1px 3px 0px, rgba(0, 0, 0, 0.13) 0px 5px 12px 0px, rgba(0, 0, 0, 0.15) 0px 0px 0.5px 0px;
+
+            @media (max-width: 700px) {
+                grid-row: 1;
+            }
+        }
+    }
+}
