+++ src/components/apps/contrast-checker/preview/types.ts
@@ -1,0 +1,7 @@
+export enum ColorDeficiencyEnum {
+    None = 'none',
+    Protanopia = 'protanopia',
+    Deutranopia = 'deutranopia',
+    Tritanopia = 'tritanopia',
+    Achromatopsia = 'achromatopsia'
+}
